Implement DBus ObjectManager approach
We want to create DBus objects for each handler. The way to make these
discoverable by clients is by implementing the ObjectManager interface.
Since this interface is complex it is easiest to resort to code generation
for it, so that's what we do here.

The current problem I'm having is that returned objects do not seem to
have the ListArgs method. I'm not sure what I'm doing wrong, but let's
commit to this approach, and look to resolve that in a later commit.