Fix UNKNOWN_OP handling
I forgot why we added UNKNOWN_OP -- it's not to indicate the target
doesn't know how to handle a SCSI opcode. That is handled by returning
CHECK CONDITION with ILLEGAL COMMAND OPERATION CODE.

No, UNKNOWN_OP is to indicate the target doesn't know how to handle a
*TCMU* opcode. Which should never happen until we ever need to add more.

So, fix this.

Also, remove the printf when we copy the sense buffer, not needed.

Signed-off-by: Andy Grover <agrover@redhat.com>