glfs: Don't get attributes that no longer exist
LIO has reduced the number of attributes that passthrough backstores
(us and pscsi) get. Now, support for these is not set via configfs, but
it's up to the handlers to decide what opcodes to support and to
set the bits (in read_capacity, or mode pages, vpds, whatever) signifying
they are supported.

Signed-off-by: Andy Grover <agrover@redhat.com>