dbus: Change API
Instead of getting arguments (aka parts of the configstring) do everything
on the cfgstring as a whole. The ConfigDesc property gives some help info
on how to format the cfgstring for a particular handler, and CheckConfig
gives the handler a chance to see the cfgstring before the TCMU dev is
created, and do some checks.

Everything is a little strange because we want to keep the DBus interface
as an optional add-on to how everything works. It can check and inform,
but we don't want to make it a required part of creating the TCMU device.

Update our two handlers to support these changes.

Signed-off-by: Andy Grover <agrover@redhat.com>