QEMU QCoW version 1 handler
Adapted from the QEMU code.  Handles qcow version 1 with compression
(read only, copy-on-write expanded) and read-only backing images in
either qcow or raw format.  Encryption is intentionally not supported,
as it's a flawed design that has been deprecated in QEMU.

My plans are to extend this to qcow2 next, but qcow was a good small
module to test porting from QEMU to tcmu-runner.

Signed-off-by: Chris Leech <cleech@redhat.com>