Refactor tcmu-runner into libtcmu.so
Move most all of the TCMU-specific parts into a library, libtcmu.so.
This should make it as easy as possible for existing apps to integrate
support for also backing TCMU devices.

Following the suggestions here:

https://git.kernel.org/cgit/linux/kernel/git/kay/libabc.git/plain/README

libtcmu provides opaque handles to the library and devices, from which
file descriptors may be obtained. The app can incorporate handling these
into its existing select/poll/epoll message loop. The libtcmu API is
defined in libtcmu.h.

However, tcmu-runner still exists on top of libtcmu, and the framework
for writing plugins for it -- where handling the poll loop is not desired
-- is still present. The tcmu-runner plugin API is defined in
tcmu-runner.h.

Helper functions for TCMU and emulating common SCSI commands are
universally usable, and are defined in libtcmu_common.h.
libtcmu_common.h.

Signed-off-by: Andy Grover <agrover@redhat.com>