libtcmu: Fix errp symbol issue with library link
Writing an application that links to libtcmu.so triggers this ld error:

    $snip/lib64/libtcmu.so: undefined reference to `errp'
    collect2: error: ld returned 1 exit status

That is because api.c references errp but there isn't a implementation.
The my_errp in libtcmu.c happens to fit the need, but just renaming this
to errp would conflict with errp in main.c and consumer.c (actually it
was a bad collision already).

Since this is an internal symbol, let's use a new name, tcmu_errp.

Signed-off-by: Fam Zheng <famz@redhat.com>