Fix spurious warning in tcmulib_processing_start/complete()
We open dev->fd in nonblocking mode, which means every time we do a read
we're going to get a -1 retval and errno=EAGAIN.

But we really don't care. We're actually only checking read()'s retval to
stop the compiler complaining. So, make the warning go away unless the
impossible happens.

Signed-off-by: Andy Grover <agrover@redhat.com>