glfs: Remove wce, tpu, tpws feature flags
WCE is now always on (if one is using the TCMU helper API for mode sense),
so now only call glfs_fdatasync() if FUA bit is set on write.

tpu and tpws (support for UNMAP and WRITE SAME) require 0xb2 (logical
block provisioning) inquiry VPD to be implemented. We have an untested
implementation for WRITE SAME, but ifdef it out until somebody implements
VPD 0xb2.

This may be a VPD we *don't* want to include in the helper API, since
different handlers may want to report something different.

Signed-off-by: Andy Grover <agrover@redhat.com>