Don't fail short/long inquirys
You cannot fail a inquiry because there is too much/litte space.
You would set the under/over flow bit at the command/transport
level.

You should set the number of bytes that would be transferred if
there was the correct amount of space in inquiry response's
ADDITIONAL LENGTH field. The initiator will then read that
and resend the inquiry with a larger buffer.