runner: add command passthrough support for generic_handle_cmd
There is no rule but I think we should only add a callout when all store drivers
should(or do) provide a backend implementation otherwise we should pass through
the commands to the specific driver which is the very(or only) one interested,
 Passthru mechanism could be analogous with system call ioctl(2) :)

Foe e.g, COMPARE_AND_WRITE should be passed through because not all store is
SCSI ware (glusterfs, sheepdog), and it can be perfectly emulated by pure read
and write, so it deserves an emulation in the generic layer to allow non-scsi 
aware backends to not care about it. This goes for ALUA stuff mentioned above 
too. Once we emulate these stuff function in the generic layer, then all the 
backend store can enjoy without even adding one line in the driver!

Signed-off-by: Liu Yuan <liuyuan@cmss.chinamobile.com>