sync tcmur handler, tcmulib_cmd and cdb naming
Originally the tcmur_handler variable was normally called
r_handler. At some point we started mixing and matching and
calling the io callbacks a store, but the other callbacks
and fields a r_handler. This kills the store naming.

We also would call the CDB's op a cmd and then sometimes
call the tcmulib_cmd a tcmulib_cmd which is a pain to write,
and then we also still had the cdb. This just has us use
cmd for tcmulib_cmd and for the op code do cdb[0].

This also fxies the handling of passthrough commands.
command_is_store_call only checks for read/write/flush
callout related commands so passthrough commands would
drop down to generic_cmd and fail. This has us check
for generic commands instead.

Signed-off-by: Mike Christie <mchristi@redhat.com>