Support for more than 1 wq thread
For compat reasons this allows handlers to specify the nr
of work queue threads. New handlers and patches for existing
ones should add config file or cfgstring support to control
this.

It also replaces aio_supported with nr_threads. If nr_threads
is set to 0, we assume the handler will handler it. We then
do not create the io wq.

Signed-off-by: Mike Christie <mchristi@redhat.com>