fix compound command buffer handling
There currently is no policy for who owns and updates the iov
and how it should be updated. It has been left to the handlers
to do what they need to track their progress.

rbd currently updates iov iov_base (when switching to sg it will not).
file and qcow do not update iov_base if the seek is not within a iov.
glfs does not update iov_base.

This results in the compound commands crashing because tcmu-runner
assumes all handlers work like rbd.

This just has compound commands store a copy of the read buffer
pointer.

Signed-off-by: Mike Christie <mchristi@redhat.com>