file optical: user preadv/preadw
Use preadv/preadw so we do not have to allocate bounce buffers
and loop over the iovec.

Fixes from
- Bryant G. Ly <bryantly@linux.vnet.ibm.com>
so fbo_read handles short preadvs and the verify code handles
short reads and also iovec handling.

- Steven Royer <seroyer@us.ibm.com>
The offset in write_verify was getting incremented during the write, but
then used again for the verify. Need to reset offset back to the
beginning of the op.

Signed-off-by: Mike Christie <mchristi@redhat.com>