runner: move tcmu_device lock to tcmur_device
The tcmu_device lock was added by the aio patches. It could be useful
for other daemons, but since they are doing their own threading/aio
they probably already handle this in a way optimized to their design.

No need to add another lock in the fast path for them, so move the
lock to the tcmur_device.

Signed-off-by: Mike Christie <mchristi@redhat.com>