runner: add possible repetitive handler registerring check.
Just encountered this when testing the old and new versions at
the same by renaming the old handlerXX.so to handlerYY.so, and
forgot to make clean or remove them. But this won't report any
warning or error message.

Maybe my way doing the test is not very wise, but this could be
a potential bug.

Signed-off-by: Xiubo Li <lixiubo@cmss.chinamobile.com>