rbd: do not reopn image from report callout
We cannot reopen the image from the report callout because
another node could have taken the lock while IO is in flight
and at the same time the network to ceph could go down.
To be safe we need that IO to fail from the blacklisting before
we can reopen the image.