tcmur: Add WRITE_SAME_16 command support
This will add a emulator for WRITE_SAME_16 command in runner. For
now the unmap option is not supported, and will be added later
when needed.

The Write Same Not Zero bit is set to one, and will be added later.

Currently we are using sync way to do the emulate to simplify the
initial patches, and the async way will be added later as commented

Signed-off-by: Xiubo Li <lixiubo@cmss.chinamobile.com>