load_module: skip loading target_core_user if already loaded
Problem:
tcmu-runner fails to start in a container as it try to load the
dependent target_core_user kernel modules and fails, this is because
loading kernel modules from container is not a valid to perform.

Instead they have to load kernel module on host manually and since
container shares their kernel between containers, so it will be made
available to container while it tries to start tcmu-runner.

Solution:
Check target_core_user module init status, if already loaded skip
loading again.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>