systemd: change KillMode from default to process
KillMode defaults to control-group.

If set to 'control-group', all remaining processes in the control group of
this unit will be killed on unit stop.

If set to 'process', only the main process itself is killed.

For more 'man systemd.kill'

Said that, with current behaviour, SIGTERM is passed to all the processes in
the control-group in case tcmu-runner terminates abnormally.

This patch changes the default control-group killMode to process.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>