runner: only support ALUA if user has explicitly set it up
For backwards compat and to avoid kernel bugs in older kernels,
this patch has runner support ALUA only if the user has
created target port groups. In the future, we can add a flag
or kernel version/feature checks if we want to have ALUA always
on and using the default group like how the kernel based LIO
does.

Signed-off-by: Mike Christie <mchristi@redhat.com>