libtcmu: make add/remove sync and pass results
This works with the kernel patch submitted here:

TODO add link/commit:

It fixes the issue where the device add and remove operations
were async, and the kernel did not get a return value, so if
runner failed to setup the device we would end up with LIO reporting
the LUN exists, but when scan related commands are sent they would
fail. In the worst case, LUN0 would be invalid and that could cause
some OSs to not scan all the LUNs.

Signed-off-by: Mike Christie <mchristi@redhat.com>