config: Add dynamic reloading support
Using the inotify, if write to the unwritable or crashed config file
forcely, the vi/vim will try to move and delete the old config file
and then recreate it again via the *.swp. When deleting the old
config file, the inotify watching handler will be removed from the
monitor too, so needs to add it back again.

Signed-off-by: Xiubo Li <lixiubo@cmss.chinamobile.com>