runner: add local target_core_user.h
It is kind of a pain to coordinate our upstream releases with
kernel headers releases. To allow users to be able to not have
to update kernel headers, and use newer runner versions, this adds a
local target_core_user.h which will stay in sync with the upstream
kernel.