runner: simplify config setup/destruction for caller
This patch does the follwoing:

1. The caller of the tcmu config setup/destroy calls does not
know the internals of the config, so it does not know if
threads are being used. This patch just has the config
code handle the details. The caller just calls setup/destroy.

2. The man page for pthread_create says if it fails thread id arg
will be in a undefined state, so we cannot just call
tcmu_cancel_config_thread since it might contain another thread's
valid id and we might end up canceling that.

Signed-off-by: Mike Christie <mchristi@redhat.com>