config: fix dynamic configure reload
The string passed to tcmu_register_option is set to
option.key, but it is a pointer to memory allocated by
tcmu_reload_config/tcmu_load_config that is later freed.

darray_append does not do a deep copy, so later if we do
darray_foreach and try to access the option->key the memory
may have been freed.

Signed-off-by: Mike Christie <mchristi@redhat.com>