libtcmu: return WCE based on device support
WCE is currently always reported as on, but some handlers
may not support it or in future patches we will want to
set it based on the emulate_write_cache bit set by the
user if the handler supports it.

This patch adds a helper to set/get the write cache
setting which is then used as the WCE bit value.

Signed-off-by: Mike Christie <mchristi@redhat.com>