Fix cluster allocation for nearly-full qcow images
The qcow handler's block allocation function originally
checked to make sure the host offset of the cluster being allocated
would not exceed the size specified in the qcow image header.
This is improper behavior, because it counted qcow metadata clusters
towards the virtual image size. This change removes the improper
check from the block allocation function and moves the bound checking
into pwritev.

Signed-off-by: Brad Warrum <bwarrum@us.ibm.com>