runner: drop command_is_generic
There is no need for command_is_generic. It just duplicates the
switch/case test. We can just run handle_generic_cmd and if it does
not handle the command drop down to the tcmur cmd handler.

handle_generic_cmd is renamed to reflect these are simple commands
that do not need to be async.

Signed-off-by: Mike Christie <mchristi@redhat.com>