runner: move target/port code to new file
The port/group code is kind of a mashup of the kernel's
alua group, target port and target port group. This was
done so we would not have to duplicate the entire kernel
state. In the next patches we will want to interact
with the target layer more to enable/disable ports and
tpgs, so this patch separates the alua port/group code
from the non alua parts.

Signed-off-by: Mike Christie <mchristi@redhat.com>