libtcmu: do not install headers and drop libtcmu stable API
When libtcmu was created it was really early in the project and
missing lots of features. It is now difficult to maintain for
possibly very little benefit espeically when it seems most
users are using the tcmu-runner daemon or really just want
low level helpers around the kernel interface.

The project will of course not change license and remain Apache
Version 2.0, so any one can just use the code in their own
project, but going forward the libtcmu interface is no longer
supported.

In the next release we will investigate a new interface.

Signed-off-by: Mike Christie <mchristi@redhat.com>