file_zbc: ZBC disk emulation handler
file_zbc.c implements a ZBC disk emulation using a file as backend
storage. The handler is compiled as handler_file_zbc.

handler_file_zbc can emulate both host aware and host managed disks.
In both cases, the LBA size, zone size and the number of conventional
zones can be configured through options in the configuration string.
The configuration, together with zone information (zone type, zone
condition, write pointer position, etc) are stored as metadata at the
beginning of the backstore file, invisibly to the user. The
configuration and state of the emulated device is thus retained
accross reboots.

The metadata in the backstore file is reinitialized (i.e. the device
configuration and state reset) if the specified backstore file does
not exist, if the metadata in the file is incorrect or if configuration
options are specified in the configuration string. Otherwise, the
metadata is reused as is, restoring the emulated device state from its
last use.

The primary goal of handler_file_zbc is kernel features development
and tests. It is especially useful for people without access to SMR
disks or on systems on which those disks cannot be used
(e.g. a laptop).

The emulated device, used either through the loopback or iscsi fabrics,
successfully passes libzbc conformance tests (using kernel 4.13-rc as
many target patches fixing various problems with sense code and sense
keys reporting are needed). Testing was also performed using f2fs
(which natively supports host managed zoned block devices) and the
dm-zoned device mapper. Additionally, the emulated device is also
currently being used for btrfs zoned block device support development
and can also be very helpful in general to test Linux block and SCSI
stack support of zoned block devices.

The blktets framework would also be another area that can benefit from
ZBC emulation.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>