libtcmu: fix name check for nl commands
If dev->dev_name is uio1 but dev_name is uio10 we would
match it and return the dev for uio1.

This would cause problems for device removal because we would
leave the cmdproc running and accessing the uio device for
the device we wanted to shutdown (we shutdown the cmdproc
for uio1 instead of uio10). The kernel would then do
uio_unregister_device and free the uio device, but runner
could still be doing uio read/poll calls and cause crashes.