libtcmu: fix possible ring buffer broken
   /* cmd_id could be different in async case */
   if (cmd->cmd_id != ent->hdr.cmd_id) {
          ent->hdr.cmd_id = cmd->cmd_id;
   }

There is no place changing and shouldn't change the cmd_id, or
this could cause the ring broken.

The old code changing the entry's cmd_id is to deal with the
async type IOs. Which will always update the first entry in the
tail, but the first entry in the tail maybe not corresponding
to the current cmd, so it in order to simplify the code here just
used a trick.

If the runner has bug like handle the complete more than once,
the TCMU will find cmd_id in idr more than once too, the ring
should be flaged as broken.

Maybe there has other bugs could cause the ring broken too.

Signed-off-by: Xiubo Li <lixiubo@cmss.chinamobile.com>