rbd: don't override the user requested max sectors
At the ceph/rbd side object sized IOs are best, but for iSCSI
it is normally going to be too large. In newer linux kernels
the initiator will try to size IOs based on the optimimal
size vpd info, and so with the default object size we are
getting 4MB IOs. Even with 10 gig ethernet we can see latencies
in the 1000s of ms and there are no improvements in throughput
with these larger IOs.

This patch just lets the user configure the IO size with the
hw_max_sectors setting.