runner/rbd: update lock state on RTPG/INQUIRY
If there is no IO in flight or sent when the lock is taken from
from a node, then rbd will not discover the lock has been taken
until it gets IO errors from being blacklisted later. This
will fire off the error handler and force retries.

To avoid that, this patch just updates our lock state when
there is a RTPG or INQUIRY. Normally this would be sent
before switching paths so we can catch the state change then
and prevent the extra retries.