runner/lib: flush device during shutdown
If remove_device is called during device removal then the kernel
will not have done the REMOVE event until all IO has stopped
and new IO will not be sent.

If remove_device is called due to the daemon/app being stopped
then the kernel will not have been notified. This has the lib
block new IO and wait for the commands in the ring to be
completed normally before removing the device.

If for some reason we hang, then we can still kill runner
and the previous patch will handle any cleaned up needed.