runner: sync alua failover type on reopens
On the initial device creation we would setup the alua settings
like failover type when we get the first command.

For reopens, we might have restarted so quickly that the initiator
does not notice and it will have not marked the path as bad, and
will not do its reinitialization. So we need to resetup the alua
failover state at device open/add time.