rbd: do not recheck size on resize event
There is not need to recheck the size on a resize event
because the calling app, gwcli/ansible, has already set
the ceph side size and then called us. This works around
the issue where if we are blacklisted the size check call
will fail and we end up with a partially updated device.