lib: check for unsupported cdb length
If tcmu_get_cdb_length detects a cdb length we do not support
we return -EINVAL and then are using that for the length.
This has us just drop the command (device will then be offlined
due to abort failures). Ideally we would fail the command, but
that would require us to add locking in the completion path in
the lib and this error just does not come up often enough in non-devel
settings for that to be worth it right now.

Signed-off-by: Mike Christie <mchristi@redhat.com>