Set correct sense upon getting write to FULL zone
According to ZBC-2, a zoned device should set INVALID FIELD IN CDB
sense when it receives a write to a full zone. Since we place the
write pointer right after the last block of a full zone, it gets
caught as unaligned write. All that needed to be done is to
set a different sense code in this situation if the zone is FULL.

Also, reduced the number of checkpatch nitpicks.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>