runner: separate connection errros from fence ones
tcmu_acquire_dev_lock should only try to reopen if the
device is fenced. If we cannot reach the cluster due to
a network issue then there is no point in trying to do
a reopen.

Also just return a new lock state instead of having the
extra TCMUR_LOCK return value.

Signed-off-by: Mike Christie <mchristi@redhat.com>