runner: get/set remote alua group info
We need to report the remote ports alua state, but the rbd handler
is not able to notify remote nodes when a lock has been taken.
This has handlers like this set the alua active optimized
alua group when the lock is taken. When a RTPG is sent we then
get this info and update our local state.

Signed-off-by: Mike Christie <mchristi@redhat.com>