runner: don't allow multiple instances of daemon to run
until now there is no check to defend on multiple runs of daemon within
the same node. This patch takes a non blocking lock on 'tcmu.lock' file,
if it succeeds only then daemon is allowed to run, if there is already a
lock on lock-file (taken by some other instance of tcmu-runner) we will exit.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>