Fix tarball release location
From jonasstein in PR
https://github.com/open-iscsi/tcmu-runner/issues/436

Tarballs are available from
https://fedorahosted.org/released/tcmu-runner/ .
in README.md
links to
https://fedoraproject.org/wiki/Infrastructure/Fedorahosted-retirement

fedorahosted.org was retired on March 1st, 2017. If you are viewing this
page, odds are it's after that date and you have been redirected here by
attempting to go to some project on fedorahosted.org. If you are a user
of a project formerly hosted at fedorahosted.org, please search for the
new home the project you are interested in likely has. If you are a
maintainer of a project formerly hosted at fedorahosted.org and need
your data to migrate, see below.