tcmu-runner: fix bug in HA about re-acquire lock
in ha, writing data to multipath device, Wait a minute, after
close the ip on node1, the write command can not be processed
on node0 by tcmu-runner rightly. Because the node1's tcmu-runner
do not drop the lock on the image, node0's tcmu-runner should
get the lock on the image after rbd return -EROFS error.

Signed-off-by: Zhaohui Yang <280965257@qq.com>