Improve generality of tcmu_dev_{err,warn,...}
Change the macros tcmu_dev_{err,warn,...} so that they can be used in a
natural way (i.e. the syntax looks like C) in more general code contexts.
In particular, a tcmu_dev_err() call such as this one:

    if (condition) {
        if (err)
                tcmu_dev_err(dev, "%s\n", err);
        else
                ret = 0;
    }

would produce an error from gcc:

    error: expected ‘}’ before ‘else’

This was due to the tcmu_dev_err macro definition enclosing its work in
braces -- that makes the semicolon at the end of the call demark an
additional "empty" C statement... which moves the "else" out of position
to be associated with the "if (err)".

    #define tcmu_dev_err(dev, ...)  {tcmu_err_message(dev, __func__, __LINE__, __VA_ARGS__);}

  (The compiler suggests it expected a '}' before the 'else', but
   that would associate the assignment with the wrong "if (condition)".)

The implemented solution is to change the macros to use the following
pattern, which demands a semicolon after a call to the macro, as would
be expected with natural C syntax:

    #define tcmu_dev_err(dev, ...)  do { tcmu_err_message(dev, __func__, __LINE__, __VA_ARGS__);} while (0)

Also, add a "missing" semicolon in one place in libtcmu.c that depended
on the old macro syntax.

Signed-off-by: David Butterfield <David.Butterfield@wdc.com>
Reviewed-by: Damien Le Moal <damien.lemoal@wdc.com>