file_zbc: correctly handle iovec seek
In the write loop in zbc_write(), tcmu_seek_in_iovec() is called to
consume the command buffer part of the written data. This can lead to
cnsuming entries in the iovec array. However, the pointer iovec is never
updated, potentially leading to a 0 length iovec use while the remaining
nr_lbas is not 0.

Fix this by properly incrementing the iovec pointer by the number of
iovec elements consumed with tcmu_seek_in_iovec().

While at it, also remove the variable iov_cnt and directly use
cmd->iov_cnt as it is used only in one place.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>