Do not return alua state transition
Linux does not handle alua state transition well starting with this
patch:

    commit 2b35865e7a290d313c3d156c0c2074b4c4ffaf52
    Author: Hannes Reinecke <hare@suse.de>
    Date:   Fri Feb 19 09:17:13 2016 +0100

        scsi_dh_alua: Recheck state on unit attention

Before this patch linux would retry for up to (cmd retries *
cmd timeout) seconds. With that patch it only retries 5 times. Because
we are not doing disk/ssd IO the retries could be used up before we even
start to take the lock. This results in the cmd being failed to the
multipath layer, the path being failed and the IO retried on another
path. This could then repeat over and over on each path.

We work around this by just returning BUSY.

Signed-off-by: Mike Christie <mchristi@redhat.com>