Fix logger thread race
If log_thread_start has removed all msgs and has returned false,
log_internal could be adding a new msg but will see thread_active
is false and not do a signal to wake log_thread_start.

To fix this we need to set thread_active to false earlier when
the rb is empty.

Signed-off-by: Mike Christie <mchristi@redhat.com>