Simplify tcmu_logbuf init
Because we set tcmu_logbuf before it is fully setup, we have to have the
finish_initialize and tcmu_logbuf_lock related code. This patch just has
log_thread_start when it has started which is the last part of the log
buf initialization.

There is no need for the locking between setup and destroy, because
the cancel call will handle it. Also the logger should be shutdown
last so there should not be any log calls in progress.

There is also no need to set a static pointer to null.

Note locking is needed around tcmu_log_dir_resetup but not the
tcmu_logbuf_lock. If we hold this while calling tcmu_err we will
deadlock because log_internal needs to grab the same lock. The next
patch will fix the locking in this path.

Signed-off-by: Mike Christie <mchristi@redhat.com>