Simplify log outputs
Drop the stdout output because in most cases it will just log to the
same place the syslog output does. We can then drop the bypass code
since we know file will be logged from the caller and syslog in the
background. The enabled code can also be dropped, because we just need
to be holding the same lock during reset that is taken when the file
output is accessed in log_internal.

Signed-off-by: Mike Christie <mchristi@redhat.com>