Remove daemon priv hack
The daemon private field on the tcmu_device is hack that should not be
exported in the new lib. This drops it and has tcmu-runner use the hm
private area for the tcmur_device then has tcmu-runner handlers use a
new tcmur_dev_set_private/tcmur_dev_get_private helpers which stores the
tcmu-runner handler private data under the tcmur_device.

Signed-off-by: Mike Christie <mchristi@redhat.com>