Build zbc/fbo with scsi.c
scsi.c is a temp file that has the old libtcmu scsi helpers. We are not
exporting these for the initial libtcmu release, so they live in that
file until we have a proper libtcmu_scsi lib. This patch builds zbc and
fbo with scsi.c and modifies that file so it does not need the libtcmu
priv header.

Signed-off-by: Mike Christie <mchristi@redhat.com>